==> demos/location-based-recommendation/generate-model/01-init-demo.ipynb <==
# Prepare initial datasets

## Install Kaggle API

!pip install kaggle --upgrade

import json

!mkdir ~/.kaggle
with open('/User/.kaggle/kaggle.json', 'w') as f:
    f.write(json.dumps(kaggle_api_token))
!chmod 600 ~/.kaggle/kaggle.json

## Create Data directory and Download zips from Kaggle

!mkdir ./data
!cd data
!kaggle competitions download -c instacart-market-basket-analysis

### Unzip & Delete

!unzip './*.zip' -d './data/'

!rm ./*.zip
